requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pyautogen==0.9.0
pydantic==2.12.4
requests==2.32.5
python-multipart==0.0.9
